+++ testlib/src/test/java/com/diffplug/spotless/markdown/FlexmarkStepTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2025 DiffPlug
+ * Copyright 2016-2026 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,10 @@
  */
 package com.diffplug.spotless.markdown;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import java.util.List;
+import java.util.Map;
 
 import org.junit.jupiter.api.Test;
 
@@ -33,6 +36,29 @@ class FlexmarkStepTest {
 				.testResource(
 						"markdown/flexmark/FlexmarkUnformatted.md",
 						"markdown/flexmark/FlexmarkFormatted.md");
+	}
+
+	@Test
+	void flexmarkOptionsRightMargin() {
+		FlexmarkConfig config = new FlexmarkConfig();
+		config.setExtensions(List.of("YamlFrontMatter"));
+		config.setFormatterOptions(Map.of("RIGHT_MARGIN", "100"));
+		StepHarness.forStep(FlexmarkStep.create(TestProvisioner.mavenCentral(), config))
+				.testResource(
+						"markdown/flexmark/FlexmarkOptionsUnformatted.md",
+						"markdown/flexmark/FlexmarkOptionsFormatted.md");
+	}
+
+	@Test
+	void flexmarkOptionsUnknownKeyFails() {
+		FlexmarkConfig config = new FlexmarkConfig();
+		config.setFormatterOptions(Map.of("NON_EXISTENT_OPTION", "value"));
+		assertThatThrownBy(() -> StepHarness.forStep(FlexmarkStep.create(TestProvisioner.mavenCentral(), config))
+				.test("text\n", "text\n"))
+				.hasRootCauseInstanceOf(IllegalArgumentException.class)
+				.hasRootCauseMessage(
+						"Unknown flexmark formatter option: no field Formatter.NON_EXISTENT_OPTION."
+								+ " See https://github.com/vsch/flexmark-java/wiki/Markdown-Formatter#options");
 	}
 
 	@Test